requirements.txt
youtube_comment_downloader
wordcloud
matplotlib
google-cloud-language
streamlit
nltk
pandas
plotly
scikit-learn
textblob
langdetect
emoji
python-dateutil
spacy>=3.0.0
tensorflow>=2.12.0
transformers>=4.30.0
torch>=2.0.0  # Required by transformers
tqdm  # Progress bars for model downloads
tf-keras
dateparser  # Universal relative date parsing with multi-language support